requirements.txt
Django==5.1.4
django-widget-tweaks==1.5.0
pillow==11.0.0
psycopg2
django-allauth==65.3.1
PyJWT==2.10.1
cryptography==44.0.0